feat(ui): SymbolDrawer replaces Today canvas and both symbol modals
One symbol surface: drawer (global) + /symbol page. Today reclaims full
width for its lists; Universes and PortfolioTable drop their local modals.
