+++ web-ui/src/components/domain/symbol/SymbolDrawer.tsx
@@ -1,11 +1,9 @@
-import { useEffect } from 'react';
 import { Link } from 'react-router-dom';
 import { ExternalLink } from 'lucide-react';
 import Drawer from '@/components/common/Drawer';
 import SymbolAnalysisContent from '@/components/domain/workspace/SymbolAnalysisContent';
 import ActionPanel from '@/components/domain/workspace/ActionPanel';
-import { syncCandidateWithFundamentals } from '@/features/screener/decisionSummary';
-import { useFundamentalSnapshotQuery } from '@/features/fundamentals/hooks';
+import { useSymbolFundamentalsSync } from '@/features/fundamentals/useSymbolFundamentalsSync';
 import { useWorkspaceStore } from '@/stores/workspaceStore';
 import { useScreenerStore } from '@/stores/screenerStore';
 import { useOpenPositions } from '@/features/portfolio/hooks';
@@ -20,34 +18,12 @@ export default function SymbolDrawer() {
   const candidate = useScreenerStore((s) =>
     ticker ? (s.lastResult?.candidates.find((c) => c.ticker === ticker) ?? null) : null
   );
-  const patchCandidate = useScreenerStore((s) => s.patchCandidate);
   const positionsQuery = useOpenPositions();
   const position = ticker
     ? (positionsQuery.data?.find((p) => p.ticker === ticker) ?? null)
     : null;
 
-  const fundamentalsQuery = useFundamentalSnapshotQuery(
-    analysisTab === 'fundamentals' ? ticker ?? undefined : undefined
-  );
-  const latestFundamentalsSnapshot = fundamentalsQuery.data;
-
-  // Fundamentals→candidate live-sync: ported VERBATIM from
-  // AnalysisCanvasPanel.tsx:33-47 (adjusted variable names to this component's).
-  useEffect(() => {
-    if (!ticker || !candidate || !latestFundamentalsSnapshot) {
-      return;
-    }
-
-    if (latestFundamentalsSnapshot.symbol.trim().toUpperCase() !== ticker.trim().toUpperCase()) {
-      return;
-    }
-
-    if (syncCandidateWithFundamentals(candidate, latestFundamentalsSnapshot) === candidate) {
-      return;
-    }
-
-    patchCandidate(ticker, (c) => syncCandidateWithFundamentals(c, latestFundamentalsSnapshot));
-  }, [latestFundamentalsSnapshot, patchCandidate, candidate, ticker]);
+  useSymbolFundamentalsSync(ticker, candidate);
 
   return (
     <Drawer
